requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
Django==5.0.4
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-grappelli==4.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
flatbuffers==25.1.24
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMatting==1.1.13
pyreadline3==3.5.4
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
referencing==0.36.2
rembg==2.0.62
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.1
scipy==1.15.1
sqlparse==0.5.0
sympy==1.13.3
text-unidecode==1.3
tifffile==2025.1.10
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0